move layout folder to component folder
