+++ components/layouts/Footer/Footer.tsx
@@ -2,12 +2,12 @@ import React from 'react';
 import { useSelector } from 'react-redux';
 import styled, { css } from 'styled-components';
 
-import Breadcrumbs from '../../components/core/Breadcrumbs/Breadcrumbs';
-import CustomLink from '../../components/core/CustomLink/CustomLink';
+import Breadcrumbs from '../../core/Breadcrumbs/Breadcrumbs';
+import CustomLink from '../../core/CustomLink/CustomLink';
 
-import { InitialState } from '../../store/reducer';
+import { InitialState } from '../../../store/reducer';
 
-import { fadeAnimation } from '../../helpers/animations';
+import { fadeAnimation } from '../../../helpers/animations';
 
 type FooterProps = {
   showVeil?: boolean;
