+++ components/layouts/Header/__tests__/header.spec.tsx
@@ -3,13 +3,13 @@ import { Provider } from 'react-redux';
 import configureStore from 'redux-mock-store';
 import { render } from '@testing-library/react';
 
-import reducer from '../../../store/reducer';
+import reducer from '../../../../store/reducer';
 
 import Header from '../Header';
 
 import { HEADER_TEXT_MOCKS } from '../__mocks__/header.mocks';
-import { CDN_URL_STATIC_DIRECTORY } from '../../../helpers/utils';
-import { renderWithTheme } from '../../../helpers/Theme';
+import { CDN_URL_STATIC_DIRECTORY } from '../../../../helpers/utils';
+import { renderWithTheme } from '../../../../helpers/Theme';
 
 const { bgImgSrc, claimTxt } = HEADER_TEXT_MOCKS;
