+++ components/layouts/__tests__/layouts.spec.tsx
@@ -4,9 +4,9 @@ import configureStore from 'redux-mock-store';
 import { render } from '@testing-library/react';
 import { ThemeProvider } from 'styled-components';
 
-import reducer from '../../store/reducer';
+import reducer from '../../../store/reducer';
 
-import { theme } from '../../helpers/Theme';
+import { theme } from '../../../helpers/Theme';
 
 import Layout from '../Layout';
