+++ components/pages/FavoritesPage/FavoritesPage.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import styled from 'styled-components';
 
-import Layout from '../../../layouts/Layout';
+import Layout from '../../layouts/Layout';
 
 import FullLoader from '../../ui/FullLoader/FullLoader';
