+++ components/pages/SearchPage/SearchPage.tsx
@@ -1,7 +1,7 @@
 import React, { forwardRef } from 'react';
 import styled from 'styled-components';
 
-import Layout from '../../../layouts/Layout';
+import Layout from '../../layouts/Layout';
 
 import Filter from '../../ui/Filter/Filter';
 import FullLoader from '../../ui/FullLoader/FullLoader';
